requirements.txt
mecab-python3==1.0.4
unidic==1.0.3
gensim==3.8.3
torch==1.9.1+cpu
sentencepiece
transformers
emoji